+++ CHANGELOG.md
@@ -1,8 +1,12 @@
-## Mongoid::Includes 1.0.0 (July 30, 2015) ##
+## Mongoid::Includes 1.0.2 (2015-10-08) ##
 
-*   Initial Version.
+*   Fix error when using `merge` or `merge!` with a criteria and `includes`.
 
-## Mongoid::Includes 1.1.0 (August 5, 2015) ##
+## Mongoid::Includes 1.0.1 (August 5, 2015) ##
 
 *   Fix error messages for polymorphic includes.
 *   Add :with option that receives the criteria that will be used to include documents.
+
+## Mongoid::Includes 1.0.0 (July 30, 2015) ##
+
+*   Initial Version.